+++ packages/core/tests/directives/arrow-function.test.ts
@@ -1,37 +1,35 @@
 import { describe, it, expect } from 'vitest';
-import * as compiler from '../src';
+import * as compiler from '../../src';
 import { ID, CLIENT, SERVER } from './example';
 
-describe('BlockStatement', () => {
+describe('ArrowFunctionExpression', () => {
   describe('client', () => {
-    it('should transform valid server block statements', async () => {
+    it('should transform valid server functions', async () => {
       const code = `
-      {
+      const example = async () => {
         'use server';
-        await doStuff();
-      }
+        return 'foo bar';
+      };
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
-    it('should skip server block statements in non-async functions', async () => {
+    it('should skip non-async server functions', async () => {
       const code = `
       const example = () => {
-        {
-          'use server';
-          doStuff();
-        }
+        'use server';
+        return 'foo bar';
       };
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
     it('should transform valid server functions with scope', async () => {
       const code = `
-      async function foo() {
+      const outer = () => {
         const value = 'foo bar';
-        {
+        const example = async () => {
           'use server';
-          await doStuff(value);
-        }
+          return value;
+        };
       }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
@@ -39,43 +37,43 @@ describe('BlockStatement', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      {
-        'use server';
-        await doStuff(value);
+      const outer = () => {
+        const example = async () => {
+          'use server';
+          return value;
+        };
       }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
   });
   describe('server', () => {
-    it('should transform valid server block statements', async () => {
+    it('should transform valid server functions', async () => {
       const code = `
-      {
+      const example = async () => {
         'use server';
-        await doStuff();
-      }
+        return 'foo bar';
+      };
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
-    it('should skip server block statements in non-async functions', async () => {
+    it('should skip non-async server functions', async () => {
       const code = `
       const example = () => {
-        {
-          'use server';
-          doStuff();
-        }
+        'use server';
+        return 'foo bar';
       };
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
     it('should transform valid server functions with scope', async () => {
       const code = `
-      async function foo() {
+      const outer = () => {
         const value = 'foo bar';
-        {
+        const example = async () => {
           'use server';
-          await doStuff(value);
-        }
+          return value;
+        };
       }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
@@ -83,9 +81,11 @@ describe('BlockStatement', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      {
-        'use server';
-        await doStuff(value);
+      const outer = () => {
+        const example = async () => {
+          'use server';
+          return value;
+        };
       }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
