+++ packages/core/tests/directives/block-statement.test.ts
@@ -1,34 +1,36 @@
 import { describe, it, expect } from 'vitest';
-import * as compiler from '../src';
+import * as compiler from '../../src';
 import { ID, CLIENT, SERVER } from './example';
 
-describe('FunctionDeclaration', () => {
+describe('BlockStatement', () => {
   describe('client', () => {
-    it('should transform valid server functions', async () => {
+    it('should transform valid server block statements', async () => {
       const code = `
-      async function example() {
+      {
         'use server';
-        return 'foo bar';
+        await doStuff();
       }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
-    it('should skip non-async server functions', async () => {
+    it('should skip server block statements in non-async functions', async () => {
       const code = `
-      function example() {
-        'use server';
-        return 'foo bar';
-      }
+      const example = () => {
+        {
+          'use server';
+          doStuff();
+        }
+      };
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
     it('should transform valid server functions with scope', async () => {
       const code = `
-      const outer = () => {
+      async function foo() {
         const value = 'foo bar';
-        async function example() {
+        {
           'use server';
-          return value;
+          await doStuff(value);
         }
       }
       `;
@@ -37,42 +39,42 @@ describe('FunctionDeclaration', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      const outer = () => {
-        async function example() {
-          'use server';
-          return value;
-        }
+      {
+        'use server';
+        await doStuff(value);
       }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
   });
   describe('server', () => {
-    it('should transform valid server functions', async () => {
+    it('should transform valid server block statements', async () => {
       const code = `
-      async function example() {
+      {
         'use server';
-        return 'foo bar';
+        await doStuff();
       }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
-    it('should skip non-async server functions', async () => {
+    it('should skip server block statements in non-async functions', async () => {
       const code = `
-      function example() {
-        'use server';
-        return 'foo bar';
-      }
+      const example = () => {
+        {
+          'use server';
+          doStuff();
+        }
+      };
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
     it('should transform valid server functions with scope', async () => {
       const code = `
-      const outer = () => {
+      async function foo() {
         const value = 'foo bar';
-        async function example() {
+        {
           'use server';
-          return value;
+          await doStuff(value);
         }
       }
       `;
@@ -81,11 +83,9 @@ describe('FunctionDeclaration', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      const outer = () => {
-        async function example() {
-          'use server';
-          return value;
-        }
+      {
+        'use server';
+        await doStuff(value);
       }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
