+++ packages/core/tests/directives/for-in-statement.test.ts
@@ -1,25 +1,25 @@
 import { describe, it, expect } from 'vitest';
-import * as compiler from '../src';
+import * as compiler from '../../src';
 import { ID, CLIENT, SERVER } from './example';
 
-describe('DoWhileStatement', () => {
+describe('ForInStatement', () => {
   describe('client', () => {
-    it('should transform valid server do-while statements', async () => {
+    it('should transform valid server for-in statements', async () => {
       const code = `
-      do {
+      for (const key in items) {
         'use server';
         await doStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
-    it('should skip server do-while statements in non-async functions', async () => {
+    it('should skip server for-in statements in non-async functions', async () => {
       const code = `
       const example = () => {
-        do {
+        for (const key in items) {
           'use server';
           doStuff();
-        } while (cond())
+        }
       };
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
@@ -28,10 +28,10 @@ describe('DoWhileStatement', () => {
       const code = `
       async function foo() {
         const value = 'foo bar';
-        do {
+        for (const key in items) {
           'use server';
           await doStuff(value);
-        } while (cond())
+        }
       }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
@@ -39,16 +39,16 @@ describe('DoWhileStatement', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      do {
+      for (const key in items) {
         'use server';
         await doStuff(value);
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
     it('should transform break statements', async () => {
       const code = `
-      do {
+      for (const key in items) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -56,13 +56,13 @@ describe('DoWhileStatement', () => {
           break;
         }
         await doMoreStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
     it('should transform continue statements', async () => {
       const code = `
-      do {
+      for (const key in items) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -70,13 +70,13 @@ describe('DoWhileStatement', () => {
           continue;
         }
         await doMoreStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
     it('should transform labeled break statements', async () => {
       const code = `
-      foo: do {
+      foo: for (const key in items) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -84,13 +84,13 @@ describe('DoWhileStatement', () => {
           break foo;
         }
         await doMoreStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
     it('should transform labeled continue statements', async () => {
       const code = `
-      foo: do {
+      foo: for (const key in items) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -98,28 +98,28 @@ describe('DoWhileStatement', () => {
           continue foo;
         }
         await doMoreStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
   });
   describe('server', () => {
-    it('should transform valid server do-while statements', async () => {
+    it('should transform valid server for-in statements', async () => {
       const code = `
-      do {
+      for (const key in items) {
         'use server';
         await doStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
-    it('should skip server do-while statements in non-async functions', async () => {
+    it('should skip server for-in statements in non-async functions', async () => {
       const code = `
       const example = () => {
-        do {
+        for (const key in items) {
           'use server';
           doStuff();
-        } while (cond())
+        }
       };
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
@@ -128,10 +128,10 @@ describe('DoWhileStatement', () => {
       const code = `
       async function foo() {
         const value = 'foo bar';
-        do {
+        for (const key in items) {
           'use server';
           await doStuff(value);
-        } while (cond())
+        }
       }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
@@ -139,16 +139,16 @@ describe('DoWhileStatement', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      do {
+      for (const key in items) {
         'use server';
         await doStuff(value);
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
     it('should transform break statements', async () => {
       const code = `
-      do {
+      for (const key in items) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -156,13 +156,13 @@ describe('DoWhileStatement', () => {
           break;
         }
         await doMoreStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
     it('should transform continue statements', async () => {
       const code = `
-      do {
+      for (const key in items) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -170,13 +170,13 @@ describe('DoWhileStatement', () => {
           continue;
         }
         await doMoreStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
     it('should transform labeled break statements', async () => {
       const code = `
-      foo: do {
+      foo: for (const key in items) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -184,13 +184,13 @@ describe('DoWhileStatement', () => {
           break foo;
         }
         await doMoreStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
     it('should transform labeled continue statements', async () => {
       const code = `
-      foo: do {
+      foo: for (const key in items) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -198,7 +198,7 @@ describe('DoWhileStatement', () => {
           continue foo;
         }
         await doMoreStuff();
-      } while (cond())
+      }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
