+++ packages/core/tests/directives/for-statement.test.ts
@@ -1,22 +1,22 @@
 import { describe, it, expect } from 'vitest';
-import * as compiler from '../src';
+import * as compiler from '../../src';
 import { ID, CLIENT, SERVER } from './example';
 
-describe('ForOfStatement', () => {
+describe('ForStatement', () => {
   describe('client', () => {
-    it('should transform valid server for-of statements', async () => {
+    it('should transform valid server for statements', async () => {
       const code = `
-      for (const item of items) {
+      for (let i = 0; i < 10; i++) {
         'use server';
         await doStuff();
       }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
-    it('should skip server for-of statements in non-async functions', async () => {
+    it('should skip server for statements in non-async functions', async () => {
       const code = `
       const example = () => {
-        for (const item of items) {
+        for (let i = 0; i < 10; i++) {
           'use server';
           doStuff();
         }
@@ -28,7 +28,7 @@ describe('ForOfStatement', () => {
       const code = `
       async function foo() {
         const value = 'foo bar';
-        for (const item of items) {
+        for (let i = 0; i < 10; i++) {
           'use server';
           await doStuff(value);
         }
@@ -39,7 +39,7 @@ describe('ForOfStatement', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      for (const item of items) {
+      for (let i = 0; i < 10; i++) {
         'use server';
         await doStuff(value);
       }
@@ -48,7 +48,7 @@ describe('ForOfStatement', () => {
     });
     it('should transform break statements', async () => {
       const code = `
-      for (const item of items) {
+      for (let i = 0; i < 10; i++) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -62,7 +62,7 @@ describe('ForOfStatement', () => {
     });
     it('should transform continue statements', async () => {
       const code = `
-      for (const item of items) {
+      for (let i = 0; i < 10; i++) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -76,7 +76,7 @@ describe('ForOfStatement', () => {
     });
     it('should transform labeled break statements', async () => {
       const code = `
-      foo: for (const item of items) {
+      foo: for (let i = 0; i < 10; i++) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -90,7 +90,7 @@ describe('ForOfStatement', () => {
     });
     it('should transform labeled continue statements', async () => {
       const code = `
-      foo: for (const item of items) {
+      foo: for (let i = 0; i < 10; i++) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -104,19 +104,19 @@ describe('ForOfStatement', () => {
     });
   });
   describe('server', () => {
-    it('should transform valid server for-of statements', async () => {
+    it('should transform valid server for statements', async () => {
       const code = `
-      for (const item of items) {
+      for (let i = 0; i < 10; i++) {
         'use server';
         await doStuff();
       }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
-    it('should skip server for-of statements in non-async functions', async () => {
+    it('should skip server for statements in non-async functions', async () => {
       const code = `
       const example = () => {
-        for (const item of items) {
+        for (let i = 0; i < 10; i++) {
           'use server';
           doStuff();
         }
@@ -128,7 +128,7 @@ describe('ForOfStatement', () => {
       const code = `
       async function foo() {
         const value = 'foo bar';
-        for (const item of items) {
+        for (let i = 0; i < 10; i++) {
           'use server';
           await doStuff(value);
         }
@@ -139,7 +139,7 @@ describe('ForOfStatement', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      for (const item of items) {
+      for (let i = 0; i < 10; i++) {
         'use server';
         await doStuff(value);
       }
@@ -148,7 +148,7 @@ describe('ForOfStatement', () => {
     });
     it('should transform break statements', async () => {
       const code = `
-      for (const item of items) {
+      for (let i = 0; i < 10; i++) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -162,7 +162,7 @@ describe('ForOfStatement', () => {
     });
     it('should transform continue statements', async () => {
       const code = `
-      for (const item of items) {
+      for (let i = 0; i < 10; i++) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -176,7 +176,7 @@ describe('ForOfStatement', () => {
     });
     it('should transform labeled break statements', async () => {
       const code = `
-      foo: for (const item of items) {
+      foo: for (let i = 0; i < 10; i++) {
         'use server';
         if (cond()) {
           await doStuff(value);
@@ -190,7 +190,7 @@ describe('ForOfStatement', () => {
     });
     it('should transform labeled continue statements', async () => {
       const code = `
-      foo: for (const item of items) {
+      foo: for (let i = 0; i < 10; i++) {
         'use server';
         if (cond()) {
           await doStuff(value);
