+++ packages/core/tests/directives/if-statement.test.ts
@@ -1,35 +1,37 @@
 import { describe, it, expect } from 'vitest';
-import * as compiler from '../src';
+import * as compiler from '../../src';
 import { ID, CLIENT, SERVER } from './example';
 
-describe('FunctionExpression', () => {
+describe('IfStatement', () => {
   describe('client', () => {
-    it('should transform valid server functions', async () => {
+    it('should transform valid server if statements', async () => {
       const code = `
-      const example = async function () {
+      if (cond()) {
         'use server';
-        return 'foo bar';
-      };
+        await doStuff();
+      }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
-    it('should skip non-async server functions', async () => {
+    it('should skip server if statements in non-async functions', async () => {
       const code = `
-      const example = function () {
-        'use server';
-        return 'foo bar';
+      const example = () => {
+        if (cond()) {
+          'use server';
+          doStuff();
+        }
       };
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
     it('should transform valid server functions with scope', async () => {
       const code = `
-      const outer = () => {
+      async function foo() {
         const value = 'foo bar';
-        const example = async function () {
+        if (cond()) {
           'use server';
-          return value;
-        };
+          await doStuff(value);
+        }
       }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
@@ -37,43 +39,43 @@ describe('FunctionExpression', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      const outer = () => {
-        const example = async function () {
-          'use server';
-          return value;
-        };
+      if (cond()) {
+        'use server';
+        await doStuff(value);
       }
       `;
       expect(await compiler.compile(code, ID, CLIENT)).toMatchSnapshot();
     });
   });
   describe('server', () => {
-    it('should transform valid server functions', async () => {
+    it('should transform valid server if statements', async () => {
       const code = `
-      const example = async function () {
+      if (cond()) {
         'use server';
-        return 'foo bar';
-      };
+        await doStuff();
+      }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
-    it('should skip non-async server functions', async () => {
+    it('should skip server if statements in non-async functions', async () => {
       const code = `
-      const example = function () {
-        'use server';
-        return 'foo bar';
+      const example = () => {
+        if (cond()) {
+          'use server';
+          doStuff();
+        }
       };
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
     });
     it('should transform valid server functions with scope', async () => {
       const code = `
-      const outer = () => {
+      async function foo() {
         const value = 'foo bar';
-        const example = async function () {
+        if (cond()) {
           'use server';
-          return value;
-        };
+          await doStuff(value);
+        }
       }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
@@ -81,11 +83,9 @@ describe('FunctionExpression', () => {
     it('should skip top-level values for scope', async () => {
       const code = `
       const value = 'foo bar';
-      const outer = () => {
-        const example = async function () {
-          'use server';
-          return value;
-        };
+      if (cond()) {
+        'use server';
+        await doStuff(value);
       }
       `;
       expect(await compiler.compile(code, ID, SERVER)).toMatchSnapshot();
