+++ packages/core/tests/directives/try-statement.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import * as compiler from '../src';
+import * as compiler from '../../src';
 import { ID, CLIENT, SERVER } from './example';
 
 describe('TryStatement', () => {
